Spray and Wait Simulation
Under construction Spray and Wait Simulation.  Nodes are successfully moving around randomly.  Using Carrier-sense MAC.  Still trying to get the routing troubleshooted.  Mechanics of it are messy, I was rushing and not thinking it through thoroughly.

If a node detects contention on the medium, it will put outgoing message it wishes to send in buffer until next turn.  It won't receive a message if more than one node are sending at the same time, collision.