Update sim00.c
It appears to be working, I think. Source gets ack back from destination that it heard message, either through relay or directly from destination.  How that happens seems consistent throughout one series of tests, but inconsistent between batches.  Not sure if this is something to do with how the random function is seeded or not.